moved calc_reflectivity to _guide_utils
